+++ src/custom/LevelGrindLayer.hpp
@@ -47,6 +47,9 @@ public:
     void onInsaneD(cocos2d::CCObject* sender);
     void onExtremeD(cocos2d::CCObject* sender);
 
+    void onDiscordBtn(cocos2d::CCObject* sender);
+    void onCreditsBtn(cocos2d::CCObject* sender);
+
     std::vector<int> difficulties;
     std::vector<int> lengths;
     std::vector<std::string> grindTypes;