+++ src/other/ReqButtonSetting.cpp
@@ -1,6 +1,6 @@
-#include "ReqButtonSetting.hpp"
+#include "../other/ReqButtonSetting.hpp"
 
-#include "ReqAction.hpp"
+#include "../other/ReqAction.hpp"
 
 #include <Geode/Geode.hpp>
 #include <Geode/binding/ButtonSprite.hpp>