+++ src/popups/UserManagePopup.hpp
@@ -8,11 +8,11 @@ using namespace geode::prelude;
 
 class UserManagePopup : public Popup {
 protected:
-    bool init(int targetAccountID, const char* username);
+    bool init(int targetAccountID, const char* username, int icon, int color1, int color2, int color3);
     async::TaskHolder<geode::utils::web::WebResponse> m_listener;
 
 public:
-    static UserManagePopup* create(int accountID, const char* username);
+    static UserManagePopup* create(int accountID, const char* username, int icon, int color1, int color2, int color3);
 
     void onDemoteBtn(CCObject* sender);
     void onPromoteBtn(CCObject* sender);
@@ -22,4 +22,9 @@ public:
 
     int m_targetAccountID;
     const char* m_username;
+
+    int m_targetIcon;
+    int m_targetColor1;
+    int m_targetColor2;
+    int m_targetColor3;
 };